+++ src/core/input.c
@@ -2,7 +2,9 @@
 #include "core/logger.h"
 #include "core/window.h"
 #include "renderer/functionRenderer.h"
+#include "ui/uiInternal.h"
 #include "utils/utilities.h"
+#include <GLFW/glfw3.h>
 
 bool isXPiLabeled = false;
 
@@ -11,6 +13,13 @@ void rih_ProcessInput(GLFWwindow *window){
     rl_LogMsg(RL_ERROR, "Window pointer passed to rih_ProcessInput is NULL.");
     return;
   }
+
+  // get mouse info
+  glfwGetCursorPos(window, &ruiState.mouseX, &ruiState.mouseY);
+  int leftMouseButton = glfwGetMouseButton(window, GLFW_MOUSE_BUTTON_LEFT);
+
+  if (leftMouseButton == GLFW_PRESS) ruiState.mouseDown = 1;
+  if (leftMouseButton == GLFW_RELEASE) ruiState.mouseDown = 0;
 
   if (glfwGetKey(window, GLFW_KEY_ESCAPE) == GLFW_PRESS){
     glfwSetWindowShouldClose(window, true);